Fixed sandbox UI. Fixed sample view controllers. Added text cell.
