+++ Playgrounds/DummyData.swift
@@ -109,6 +109,7 @@ public struct DummyData {
             plot(t1.elements, siny1.elements)
             scatter(t2.elements, siny2.elements)
         }
-        print(t1)
+        nwprint(t1)
+        nwprint("aaaa\naaaa\naaaa\naaaa\naaaa\naaaa\naaaa\n")
     }
 }
