+++ Playgrounds/RenderTableViewCell.swift
@@ -1,0 +1,59 @@
+//
+//  RenderTableViewCell.swift
+//  sandbox
+//
+//  Created by sonson on 2017/04/07.
+//  Copyright © 2017年 sonson. All rights reserved.
+//
+
+import UIKit
+
+internal class RenderTableViewCell: UITableViewCell {
+    
+    var renderer: Renderer? {
+        willSet {
+            self.renderImageView.image = nil
+            self.renderedImageSize = .zero
+        }
+    }
+    
+    private let renderImageView = UIImageView(frame: .zero)
+    
+    override init(style: UITableViewCellStyle, reuseIdentifier: String?) {
+        super.init(style: .default, reuseIdentifier: reuseIdentifier)
+        print("RenderTableViewCell init")
+        self.separatorInset = .zero
+        self.selectionStyle = .none
+        renderImageView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
+        renderImageView.contentMode = .scaleAspectFit
+        self.contentView.addSubview(renderImageView)
+        renderImageView.frame = self.contentView.bounds
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    deinit {
+        print("RenderTableViewCell deinit")
+    }
+    
+    private var renderedImageSize = CGSize.zero
+    
+    func updateImageViewIfNeeded() {
+        print("rendering bounds: \(self.contentView.bounds)")
+        if renderedImageSize == self.contentView.bounds.size &&
+            self.renderImageView.image != nil {
+            // already rendered
+            return
+        }
+        updateImageView()
+    }
+    
+    private func updateImageView() {
+        guard let renderer = self.renderer else { return }
+        let image = renderer.renderToImage(size: self.contentView.bounds.size)
+        self.renderImageView.image = image
+        renderedImageSize = self.contentView.bounds.size
+    }
+}
