+++ Playgrounds/RenderTableViewController.swift
@@ -11,58 +11,7 @@
 #if os(iOS)
 import UIKit
 
-private class RenderTableViewCell: UITableViewCell {
-    
-    var renderer: Renderer? {
-        willSet {
-            self.renderImageView.image = nil
-            self.renderedImageSize = .zero
-        }
-    }
-    
-    private let renderImageView = UIImageView(frame: .zero)
-    
-    override init(style: UITableViewCellStyle, reuseIdentifier: String?) {
-        super.init(style: .default, reuseIdentifier: reuseIdentifier)
-        print("RenderTableViewCell init")
-        self.separatorInset = .zero
-        self.selectionStyle = .none
-        renderImageView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
-        renderImageView.contentMode = .scaleAspectFit
-        self.contentView.addSubview(renderImageView)
-        renderImageView.frame = self.contentView.bounds
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    deinit {
-        print("RenderTableViewCell deinit")
-    }
-    
-    private var renderedImageSize = CGSize.zero
-    
-    func updateImageViewIfNeeded() {
-        print("rendering bounds: \(self.contentView.bounds)")
-        if renderedImageSize == self.contentView.bounds.size &&
-            self.renderImageView.image != nil {
-            // already rendered
-            return
-        }
-        updateImageView()
-    }
-    
-    private func updateImageView() {
-        guard let renderer = self.renderer else { return }
-        let image = renderer.renderToImage(size: self.contentView.bounds.size)
-        self.renderImageView.image = image
-        renderedImageSize = self.contentView.bounds.size
-    }
-}
-
 public class RenderTableViewController: UITableViewController {
-    private let CellIdentifier = "Cell"
     
     var renderers: [Renderer] = [] {
         didSet {
@@ -87,7 +36,8 @@ public class RenderTableViewController: UITableViewController {
         tableView.contentInset = .zero
         tableView.separatorStyle = .none
         
-        tableView.register(RenderTableViewCell.self, forCellReuseIdentifier: CellIdentifier)
+        tableView.register(RenderTableViewCell.self, forCellReuseIdentifier: "RenderTableViewCell")
+        tableView.register(TextTableViewCell.self, forCellReuseIdentifier: "TextTableViewCell")
     }
     
     public func append(renderer: Renderer) {
@@ -111,9 +61,14 @@ public class RenderTableViewController: UITableViewController {
     }
     
     public override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: CellIdentifier, for: indexPath)
-        if let cell = cell as? RenderTableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: renderers[indexPath.row].cellIdentifier, for: indexPath)
+        switch (cell, renderers[indexPath.row]) {
+        case (let cell as RenderTableViewCell, _):
             cell.renderer = renderers[indexPath.row]
+        case (let cell as TextTableViewCell, let renderer as TextRenderer):
+            cell.textView.attributedString = renderer.attributedString
+        default:
+            do {}
         }
         return cell
     }
