+++ Playgrounds/TextTableViewCell.swift
@@ -1,0 +1,40 @@
+//
+//  TextTableViewCell.swift
+//  sandbox
+//
+//  Created by sonson on 2017/04/07.
+//  Copyright © 2017年 sonson. All rights reserved.
+//
+
+import UIKit
+
+internal class TextTableViewCell: UITableViewCell {
+    let textView = UZTextView(frame: .zero)
+
+    override func awakeFromNib() {
+        super.awakeFromNib()
+        // Initialization code
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        super.init(coder: aDecoder)
+    }
+    
+    override init(style: UITableViewCellStyle, reuseIdentifier: String?) {
+        super.init(style: style, reuseIdentifier: reuseIdentifier)
+        self.separatorInset = .zero
+        self.selectionStyle = .none
+        textView.backgroundColor = .white
+        textView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
+        textView.contentMode = .scaleAspectFit
+        self.contentView.addSubview(textView)
+        textView.frame = self.contentView.bounds
+    }
+
+    override func setSelected(_ selected: Bool, animated: Bool) {
+        super.setSelected(selected, animated: animated)
+
+        // Configure the view for the selected state
+    }
+
+}
