+++ Playgrounds/sandbox/sandbox/RendererDebugViewController.swift
@@ -18,25 +18,10 @@ class RendererDebugViewController: UIViewController {
     }
     
     required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
+        super.init(coder: aDecoder)
     }
     
     override func viewDidLoad() {
-        func makeRenderer() -> ChartRenderer {
-            let points1 = DummyData.points1()
-            let points2 = DummyData.points2()
-            
-            var chart = Chart()
-            chart.elements = [
-                .line(LineGraph(points: points1)),
-                .line(LineGraph(points: points2))
-            ]
-            chart.computeViewport()
-            
-            return ChartRenderer(chart: chart)
-        }
-        
-        renderer = makeRenderer()
         
     }
