--- Playgrounds/sandbox/sandbox/RendererDebugViewController.xib
@@ -1,78 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="11762" systemVersion="16C67" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
-    <device id="ipad9_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="11757"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="RendererDebugViewController" customModule="sandbox" customModuleProvider="target">
-            <connections>
-                <outlet property="imageView" destination="gzP-hT-Z59" id="Xkg-jq-KJJ"/>
-                <outlet property="view" destination="iN0-l3-epB" id="zuz-nv-YmR"/>
-            </connections>
-        </placeholder>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB">
-            <rect key="frame" x="0.0" y="0.0" width="768" height="1024"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="7Fb-bY-91q">
-                    <rect key="frame" x="20" y="50" width="42" height="21"/>
-                    <constraints>
-                        <constraint firstAttribute="width" constant="42" id="3Ga-2W-vgP"/>
-                        <constraint firstAttribute="height" constant="21" id="QGA-us-mrp"/>
-                    </constraints>
-                    <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                    <nil key="textColor"/>
-                    <nil key="highlightedColor"/>
-                </label>
-                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="eVr-pf-JHI">
-                    <rect key="frame" x="105" y="45" width="60" height="30"/>
-                    <constraints>
-                        <constraint firstAttribute="height" constant="30" id="NOo-Ws-Ffy"/>
-                        <constraint firstAttribute="width" constant="60" id="vSp-i2-SWx"/>
-                    </constraints>
-                    <state key="normal" title="Render"/>
-                    <connections>
-                        <action selector="onRenderButtonWithSender:" destination="-1" eventType="touchUpInside" id="Tcs-hV-JqL"/>
-                    </connections>
-                </button>
-                <imageView userInteractionEnabled="NO" contentMode="scaleToFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="gzP-hT-Z59">
-                    <rect key="frame" x="20" y="83" width="728" height="600"/>
-                    <color key="backgroundColor" red="0.92851360103626945" green="0.91807155525744699" blue="0.64710142105149004" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-                    <constraints>
-                        <constraint firstAttribute="height" constant="600" id="4tl-04-jeB"/>
-                    </constraints>
-                </imageView>
-                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="xw6-fJ-a3q">
-                    <rect key="frame" x="209" y="45" width="60" height="30"/>
-                    <constraints>
-                        <constraint firstAttribute="width" constant="60" id="6Ze-lW-aks"/>
-                        <constraint firstAttribute="height" constant="30" id="FOp-ee-lva"/>
-                    </constraints>
-                    <state key="normal" title="Clear"/>
-                    <connections>
-                        <action selector="onClearButton" destination="-1" eventType="touchUpInside" id="hHk-hU-eHZ"/>
-                    </connections>
-                </button>
-            </subviews>
-            <color key="backgroundColor" red="0.68594512617952641" green="0.92851360103626945" blue="0.71137366029529925" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-            <constraints>
-                <constraint firstItem="7Fb-bY-91q" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" constant="20" id="AII-fa-FSm"/>
-                <constraint firstItem="xw6-fJ-a3q" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" constant="45" id="DYd-3V-IBP"/>
-                <constraint firstAttribute="trailing" secondItem="gzP-hT-Z59" secondAttribute="trailing" constant="20" id="Fqz-Ig-61p"/>
-                <constraint firstItem="gzP-hT-Z59" firstAttribute="top" secondItem="eVr-pf-JHI" secondAttribute="bottom" constant="8" id="MRl-XU-xdI"/>
-                <constraint firstItem="7Fb-bY-91q" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" constant="50" id="Mjy-m2-2I3"/>
-                <constraint firstItem="xw6-fJ-a3q" firstAttribute="leading" secondItem="eVr-pf-JHI" secondAttribute="trailing" constant="44" id="Ps0-sV-YGe"/>
-                <constraint firstItem="gzP-hT-Z59" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" constant="20" id="Z4l-yi-A2j"/>
-                <constraint firstItem="eVr-pf-JHI" firstAttribute="leading" secondItem="7Fb-bY-91q" secondAttribute="trailing" constant="43" id="ZId-CZ-uMx"/>
-                <constraint firstItem="eVr-pf-JHI" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" constant="45" id="oIz-yF-hsP"/>
-            </constraints>
-            <point key="canvasLocation" x="41" y="59"/>
-        </view>
-    </objects>
-</document>
